Address Sanghyeok's comments: Refactor containers and GFlowNet modules
- Update import statements in containers and GFlowNet modules
- Modify log_probs handling in Trajectories class
- Add warning for recalculating log probabilities
- Use type casting in TBGFlowNet loss calculation
- Improve method signatures and docstrings
